DAY-06 Add solution part 2

+++ src/main/kotlin/AdventDay.kt
@@ -3,8 +3,8 @@ abstract class AdventDay(fileName: String) {
     protected val puzzle: List<String> = readFileFromClasspath(fileName)
         .split("\n")
 
-    abstract fun part1(): Int
-    abstract fun part2(): Int
+    abstract fun part1(): Long
+    abstract fun part2(): Long
     abstract fun printResult()
 
     private fun readFileFromClasspath(path: String): String =
